refactoring to move common modules into common lib
