--- projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/exceptions/BusinessInteractionNotFoundException.kt
@@ -1,4 +1,0 @@
-package uk.gov.justice.digital.hmpps.exceptions
-
-class BusinessInteractionNotFoundException(code: String) :
-    RuntimeException("No Business Interaction found with code $code")
